+++ chainpack/__init__.py
@@ -8,7 +8,9 @@ from . import (
     rpcvalue,
 )
 from .rpcclient import RpcClient
+from .rpcserver import RpcServer
 from .rpcmessage import RpcMessage
+from .rpcprotocol import RpcProtocol
 
 __all__ = [
     "chainpack",
@@ -19,5 +21,7 @@ __all__ = [
     "rpcmessage",
     "rpcvalue",
     "RpcClient",
+    "RpcServer",
     "RpcMessage",
+    "RpcProtocol",
 ]