+++ chainpack/rpcclient.py
@@ -1,14 +1,17 @@
+"""RPC connection, that includes client and specific server connection."""
+from __future__ import annotations
 import asyncio
+import enum
 import logging
 import typing
-from enum import Enum
 
 from .chainpack import ChainPack, ChainPackReader, ChainPackWriter
 from .cpcontext import UnpackContext
 from .cpon import Cpon, CponReader, CponWriter
 from .rpcmessage import RpcMessage
+from .rpcprotocol import RpcProtocol
 
-_logger = logging.getLogger("RpcClient")
+logger = logging.getLogger(__name__)
 
 
 def get_next_rpc_request_id():
@@ -17,6 +20,15 @@ def get_next_rpc_request_id():
 
 
 class RpcClient:
+    """RPC connection to some SHV peer.
+
+    You most likely want to use `connect` or `connect_device` class methods
+    instead of initializing this class directly.
+
+    :param reader: Reader for the connection to the SHV RPC server.
+    :param writer: Writer for the connection to the SHV RPC server.
+    """
+
     class MethodCallError(Exception):
         def __init__(self, error):
             self.error = error
@@ -24,69 +36,138 @@ class RpcClient:
         def __str__(self):
             return CponWriter.pack(self.error).decode()
 
-    class State(Enum):
-        Closed = 1
-        Connecting = 2
-        Connected = 3
-        LoggedIn = 4
-        ConnectionError = 5
-
-    class LoginType(Enum):
-        Plain = "PLAIN"
-        Sha1 = "SHA1"
+    class LoginType(enum.Enum):
+        """Enum specifying which login type should be used."""
+        PLAIN = "PLAIN"
+        """Plain login format should be used."""
+        SHA1 = "SHA1"
+        """Use hash algorithm SHA1 (preferred and common default)."""
+        NONE = None
+        """No login type thus perform no login."""
 
     lastRequestId = 0
 
-    def __init__(self):
-        self.state = RpcClient.State.Closed
-        self.readData = bytearray(0)
-        self.reader = None
-        self.writer = None
-
-    async def connect(
+    def __init__(
         self,
-        host,
-        port=3755,
-        user=None,
-        password=None,
-        login_type: LoginType = LoginType.Sha1,
-        device_id=None,
-        mount_point=None,
+        reader: asyncio.StreamReader,
+        writer: asyncio.StreamWriter,
     ):
-        _logger.debug("connecting to: {}:{}".format(host, port))
-        self.state = RpcClient.State.Connecting
-        self.reader, self.writer = await asyncio.open_connection(host, port)
-        self.state = RpcClient.State.Connected
-        _logger.debug("TCP CONNECTED")
-        await self.call_shv_method(None, "hello")
-        await self.read_rpc_message()
+        self.read_data = bytearray(0)
+        self.reader = reader
+        self.writer = writer
 
+    @classmethod
+    async def connect(
+        cls,
+        host: str | None,
+        port: int = 3755,
+        protocol: RpcProtocol = RpcProtocol.TCP,
+        user: str | None = None,
+        password: str | None = None,
+        login_type: LoginType = LoginType.SHA1,
+        login_options: dict[str, typing.Any] = {"idleWatchDogTimeOut": 0},
+    ) -> RpcClient:
+        """Connect and login to the SHV RPC server.
+
+        :param host: IP/Hostname (TCP) or socket path (LOCKAL_SOCKET)
+        :param port: Port (TCP) to connect to
+        :param protocol: Protocol used to connect to the server
+        :param user: User name used to login
+        :param password: Password used to login
+        :param login_type: Type of the login to be used
+        :param login_options: Options sent with login
+        :returns: Connected and logged in SHV RPC client handle
+        """
+        if host is None:
+            if protocol == RpcProtocol.TCP:
+                host = "localhost"
+            elif protocol == RpcProtocol.LOCAL_SOCKET:
+                host = "shv.sock"
+            else:
+                raise RuntimeError(f"Invalid protocol: {protocol}")
+        logger.debug("Connecting to: %s:%d", host, port)
+        if protocol == RpcProtocol.TCP:
+            reader, writer = await asyncio.open_connection(host, port)
+        elif protocol == RpcProtocol.LOCAL_SOCKET:
+            reader, writer = await asyncio.open_unix_connection(host)
+        client = cls(reader, writer)
+        logger.debug("%s CONNECTED", str(protocol))
+
+        if login_type is cls.LoginType.NONE:
+            return client
+
+        await client.call_shv_method(None, "hello")
+        await client.read_rpc_message()
         params = {
             "login": {"password": password, "type": login_type.value, "user": user},
-            "options": {
-                # "device": {
-                #     "deviceId": "dev-id"
-                #     "mountPoint": "test/agent1"
-                # },
-                "idleWatchDogTimeOut": 0
-            },
+            "options": login_options,
         }
-        if device_id is not None:
-            params["options"]["device"] = {"deviceId": device_id}
-        elif mount_point is not None:
-            params["options"]["device"] = {"mountPoint": mount_point}
-        _logger.debug("LOGGING IN")
-        await self.call_shv_method(None, "login", params)
-        await self.read_rpc_message()
-        _logger.debug("LOGGED IN")
-        self.state = RpcClient.State.LoggedIn
+        logger.debug("LOGGING IN")
+        await client.call_shv_method(None, "login", params)
+        await client.read_rpc_message()
+        logger.debug("LOGGED IN")
+        return client
 
-    async def call_shv_method(self, shv_path, method, params=None):
-        await self.call_shv_method_with_id(
-            get_next_rpc_request_id(), shv_path, method, params
+    @classmethod
+    async def connect_device(
+        cls,
+        device_id: int,
+        mount_point: str,
+        host: str | None,
+        port: int = 3755,
+        protocol: RpcProtocol = RpcProtocol.TCP,
+        user: str | None = None,
+        password: str | None = None,
+        login_type: LoginType = LoginType.SHA1,
+    ) -> RpcClient:
+        """Connect and login to the SHV RPC server when being device.
+
+        This is variant of `connect` class method that should be used when
+        connecting a new device to the broker.
+
+        The parameters are the same as in case of `connect` with exception of
+        the documented ones.
+
+        :param device_id: Identifier of this device
+        :param mount_point: Path the device should be mounted to
+        """
+        return cls.connect(
+            host,
+            port,
+            protocol,
+            user,
+            password,
+            login_type,
+            {
+                "device": {
+                    "deviceId": device_id,
+                    "mountPoint": mount_point,
+                },
+            },
         )
 
-    async def call_shv_method_with_id(self, req_id, shv_path, method, params=None):
+    async def call_shv_method(self, shv_path, method, params=None) -> int:
+        """Call the given SHV method on given path.
+
+        :param shv_path: Path to the node with requested method
+        :param method: Name of the method to call
+        :param params: Parameters passed to the method
+        :returns: assigned request ID you can use to identify the response
+        """
+        rid = get_next_rpc_request_id()
+        await self.call_shv_method_with_id(rid, shv_path, method, params)
+        return rid
+
+    async def call_shv_method_with_id(
+        self, req_id, shv_path, method, params=None
+    ) -> None:
+        """Call the given SHV method on given path with specific request ID.
+
+        :param req_id: Request ID used to submit the method call with.
+        :param shv_path: Path to the node with requested method
+        :param method: Name of the method to call
+        :param params: Parameters passed to the method
+        """
         msg = RpcMessage()
         msg.set_shv_path(shv_path)
         msg.set_method(method)
@@ -95,43 +176,43 @@ class RpcClient:
 
         await self.send_rpc_message(msg)
 
-    async def send_rpc_message(self, msg):
+    async def send_rpc_message(self, msg: RpcMessage) -> None:
+        """Send the given SHV RPC Message.
+
+        :param msg: Message to be sent
+        """
         data = msg.to_chainpack()
-        _logger.debug("<== SND: {}".format(msg.to_string()))
+        logger.debug("<== SND: {}".format(msg.to_string()))
 
-        wr = ChainPackWriter()
-        wr.write_uint_data(len(data) + 1)
-        self.writer.write(wr.ctx.data_bytes())
-
-        proto = bytearray(1)
-        proto[0] = ChainPack.ProtocolType
-        self.writer.write(proto)
-
+        writter = ChainPackWriter()
+        writter.write_uint_data(len(data) + 1)
+        self.writer.write(writter.ctx.data_bytes())
+        self.writer.write(ChainPack.ProtocolType.to_bytes(1, "big"))
         self.writer.write(data)
         await self.writer.drain()
 
     def _get_rpc_msg(self):
-        if len(self.readData) < 6:
+        if len(self.read_data) < 6:
             return None
         try:
-            rd = ChainPackReader(self.readData)
+            rd = ChainPackReader(self.read_data)
             size = rd.read_uint_data()
             packet_len = size + rd.ctx.index
         except UnpackContext.BufferUnderflow:
             return None
-        if packet_len > len(self.readData):
+        if packet_len > len(self.read_data):
             return None
         proto = rd.ctx.get_byte()
         if proto == Cpon.ProtocolType:
             rd = CponReader(rd.ctx)
         rpc_val = rd.read()
-        self.readData = self.readData[packet_len:]
+        self.read_data = self.read_data[packet_len:]
         return RpcMessage(rpc_val)
 
     async def read_rpc_message(
         self, throw_error: bool = True
     ) -> typing.Optional[RpcMessage]:
-        """Read next received RPC message.
+        """Read next received RPC message or wait for next to be received.
 
         :param throw_error: If MethodCallError should be raised or not.
         :returns: Next RPC message is returned or `None` in case of EOF.
@@ -140,10 +221,10 @@ class RpcClient:
         while not self.reader.at_eof():
             msg = self._get_rpc_msg()
             if msg:
-                _logger.debug("==> REC: {}".format(msg.to_string()))
+                logger.debug("==> REC: {}".format(msg.to_string()))
                 if throw_error and msg.error():
                     raise RpcClient.MethodCallError(msg.error())
                 return msg
             data = await self.reader.read(1024)
-            self.readData += data
+            self.read_data += data
         return None